Added news and article pages
News page has been added as well as the article page, which is the specific article the user clicks on.

12 actual movies have been added to the database and with that 12 movie posters.

Changed movie table in the Database, so the the data value for movie description is now text instead of varchar.

All previous made pages accessible for the user have been changed, so all pages now have news added to the nav bar.