Added News articles & showtimes to DB
Created 6 news articles and added them into the database, so that I have data to work with when creating the homepage and news functions.

Created 720 showtimes and added them to the Database. Covering the next 3 month, so there will be data every day until the exam is over.

Changed how showtimes are ordered, so it only looks at date ascending. If there is multiple with the same date (which there is), it prioritizes by the order it was inserted into the DB.