Unfixed my fix...
The Seats table WAS fine (I think?). According to my RDM and my visualisation of the seats table the original table worked perfectly fine as long as the data can be displayed in grids in the way it was.

There for the table has been changed back and Seats Data for both halls has been added.