Created Booking function
Changed ShowtimeDisplay class to Select Showtime_id as this is needed to create a proper function for booking.

Changed the URL for showtimes to include id, so that The right bookings can be made.

Created BookingDisplay which get needed data from the specific movie, hall, showtime and seats.

Added functions to booking page to call the data and seats. Added modal to confirm the users choices.

Reinforced the rules sat in place for seat selection.